requirements.txt
# langchain_openai 
# langchain_core
# python-dotenv
# streamlit
# langchain_community
# langserve
# fastapi
# uvicorn
# sse_starlette
# bs4
# pypdf
# chromadb
# faiss-cpu
# groq
# cassio
# beautifulsoup4
# langchain-groq
# wikipedia
# arxiv
# langchainhub
# sentence_transformers
# PyPDF2
# langchain-objectbox
# langchain
streamlit
python-dotenv
langchain
langchain_community
faiss-cpu
langchain-groq
#sentence_transformers
pinecone-client
langchain-pinecone